requirements.txt
asgiref==3.5.2
black==22.6.0
click==8.1.3
colorama==0.4.5
dj-database-url==1.0.0
Django==4.0.6
django-cors-headers==3.13.0
django-heroku==0.3.1
djangorestframework==3.13.1
isort==5.10.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
psycopg2==2.9.4
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.3.0
tzdata==2022.1
whitenoise==6.2.0
